Fixed an issue with XML serialization.
The inner-loop in to_xml() which creates the XML element tree (etree) was not adding the type property
on XML nodes that are the serializations of bundles. As a result, tastypie cannot consume its own
XML output.

I've added a element.set("type", "hash") call so that we get the models instances back as XML
hashes.